# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_cv_import_api_onboarding_cv_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_cv_import_api_onboarding_cv_import_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnboardingFinaliseRequest": {
        "description": "Web wizard's fully-typed payload.\n\nStructured stages are pre-validated by the frontend. Motivation and\ndeal-breaker prose is split deterministically into retrievable preferences.\nWriting-sample onboarding was removed; V2 style guidance comes from\nexplicit voice/persona settings and the browser companion feedback loop.",
        "properties": {
          "base_location": {
            "title": "Base Location",
            "type": "string"
          },
          "career_narrative": {
            "default": "",
            "title": "Career Narrative",
            "type": "string"
          },
          "current_employment": {
            "enum": [
              "EMPLOYED",
              "UNEMPLOYED",
              "NOTICE_PERIOD"
            ],
            "title": "Current Employment",
            "type": "string"
          },
          "deal_breakers_text": {
            "default": "",
            "title": "Deal Breakers Text",
            "type": "string"
          },
          "good_role_signals_text": {
            "default": "",
            "title": "Good Role Signals Text",
            "type": "string"
          },
          "life_constraints": {
            "items": {
              "type": "string"
            },
            "title": "Life Constraints",
            "type": "array"
          },
          "motivations_text": {
            "default": "",
            "title": "Motivations Text",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "salary_floor": {
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Salary Floor",
            "type": "integer"
          },
          "salary_target": {
            "anyOf": [
              {
                "maximum": 10000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Target"
          },
          "search_duration_months": {
            "anyOf": [
              {
                "maximum": 240.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Duration Months"
          },
          "user_type": {
            "enum": [
              "visa_holder",
              "uk_resident"
            ],
            "title": "User Type",
            "type": "string"
          },
          "visa_expiry": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visa Expiry"
          },
          "visa_route": {
            "anyOf": [
              {
                "enum": [
                  "graduate",
                  "skilled_worker",
                  "dependant",
                  "student",
                  "global_talent",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visa Route"
          }
        },
        "required": [
          "name",
          "user_type",
          "base_location",
          "salary_floor",
          "current_employment"
        ],
        "title": "OnboardingFinaliseRequest",
        "type": "object"
      },
      "OnboardingFinaliseResponse": {
        "properties": {
          "career_entries_written": {
            "title": "Career Entries Written",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "career_entries_written"
        ],
        "title": "OnboardingFinaliseResponse",
        "type": "object"
      },
      "OnboardingParseRequest": {
        "description": "Helper for the wizard when it wants to show the user a parsed\nsummary before accepting their free-text answer. Finalise does not\ncall this; it is exposed only for future per-stage clarification UX.",
        "properties": {
          "stage": {
            "enum": [
              "career",
              "motivations",
              "deal_breakers",
              "money",
              "visa",
              "life"
            ],
            "title": "Stage",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "text"
        ],
        "title": "OnboardingParseRequest",
        "type": "object"
      },
      "OutcomeRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outcome": {
            "enum": [
              "applied",
              "no_response",
              "rejected_screen",
              "rejected_interview",
              "rejected_offer",
              "offer_received",
              "offer_accepted",
              "offer_declined"
            ],
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "OutcomeRequest",
        "type": "object"
      },
      "OutcomeResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "outcome": {
            "enum": [
              "applied",
              "no_response",
              "rejected_screen",
              "rejected_interview",
              "rejected_offer",
              "offer_received",
              "offer_accepted",
              "offer_declined"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "outcome"
        ],
        "title": "OutcomeResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Web surface for the AskPicky job-search assistant.",
    "title": "AskPicky API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/onboarding/cv_import": {
      "post": {
        "description": "Extract structured data from an uploaded CV (PDF / DOCX / TXT).\n\nSingle DeepSeek V4 Flash call (~5s) \u2014 returns the full CVImport shape with\nname, location, contact email, role rows + bullets, education,\nprojects, skills, professional_summary, and the chronological\nnarrative bio.\n\nStateless \u2014 no session or profile is written here; the wizard's\nlocalStorage draft holds the result until the user clicks Finish.",
        "operationId": "cv_import_api_onboarding_cv_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_cv_import_api_onboarding_cv_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cv Import Api Onboarding Cv Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cv Import"
      }
    },
    "/api/onboarding/finalise": {
      "post": {
        "description": "Write UserProfile + CareerEntries.\n\nNo LLM parsers run here. All writes share the same `now` timestamp so\ndownstream queries can correlate them.",
        "operationId": "finalise_api_onboarding_finalise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingFinaliseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingFinaliseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalise"
      }
    },
    "/api/onboarding/parse": {
      "post": {
        "description": "Run the onboarding parser for a single free-text stage.\n\nReturns the raw ParseResult shape (status + fields + follow_up).\nThe wizard can use this to show a parsed summary before the user\nmoves on, or display a clarification question on\nneeds_clarification. Wave 9 wizard calls /finalise only; this\nendpoint is exposed for future richer UX.",
        "operationId": "parse_api_onboarding_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Parse Api Onboarding Parse Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse"
      }
    },
...
    "/api/sessions/{session_id}/outcome": {
      "post": {
        "description": "Record the terminal outcome of an application.\n\nOwnership-gated: 404 covers both \"session not found\" and \"not yours\"\nso an attacker cannot enumerate session ids.",
        "operationId": "record_outcome_api_sessions__session_id__outcome_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Outcome"
      }
    },
...
  }
}